token

+++ frontend/src/pages/Signin.tsx
@@ -1,16 +1,17 @@
 import { useNavigate } from "react-router-dom"
 import Navbar from "../components/Navbar"
 import axios from "axios"
-import { useState } from "react"
+import { useEffect, useState } from "react"
 
 export default function Signin() {
     const navigate = useNavigate()
-    let token = window.localStorage.getItem("token")
-
-    if (token) {
-        navigate("/home");
-    }
-
+    let token: string;
+    useEffect(() => {
+        token = window.localStorage.getItem("token")?.toString() || "Unknown"
+        if (token && token != "Unknown") {
+            navigate("/home");
+        }
+    }, [])
     const [inputs, setInputs] = useState({
         username: "",
         password: ""
@@ -19,15 +20,15 @@ export default function Signin() {
     return (
         <div className="flex flex-col bg-black min-h-screen">
             <Navbar />
-            <div className="flex items-center h-screen justify-center m-4 mb-40">
+            <div className="flex items-center h-full mt-40 justify-center m-4">
                 <div className="p-6 w-[350px] flex flex-col justify-center text-white border border-gray-800 shadow-sm font-mono rounded-lg ">
                     <div className="flex justify-between">
                         <div className="font-semibold text-2xl">
                             Signin
                         </div>
                     </div>
-                    <label className="py-1" htmlFor="Name">Name</label>
-                    <input placeholder="Enter Your Name" id="Name" type="text" onChange={(e) => {
+                    <label className="py-1" htmlFor="Username">Username</label>
+                    <input placeholder="Enter Your Username" id="Username" type="text" onChange={(e) => {
                         setInputs({
                             ...inputs,
                             username: e.target.value
@@ -35,7 +36,7 @@ export default function Signin() {
                     }}
                         className="p-2 mb-3 border border-slate-800  bg-background rounded-lg" />
                     <label className="py-1" htmlFor="password">Password</label>
-                    <input placeholder="Enter Your password" id="password" type="text"  onChange={(e) => {
+                    <input placeholder="Enter Your password" id="password" type="text" onChange={(e) => {
                         setInputs({
                             ...inputs,
                             password: e.target.value
@@ -43,7 +44,7 @@ export default function Signin() {
                     }}
                         className="p-2 mb-3 border border-slate-800  bg-background rounded-lg" />
                     <div className="flex justify-center mt-2">
-                        <button className="border border-slate-800 w-full py-2 bg-background rounded-4xl"
+                        <button className="border border-slate-800 w-full py-2 bg-background rounded-md"
                             onClick={async () => {
                                 if (!inputs.password || !inputs.username) {
                                     window.alert("Please fill in all fields.");
@@ -54,8 +55,17 @@ export default function Signin() {
                                         const response = await axios.post("http://localhost:3000/signin", {
                                             inputs
                                         })
-                                        window.localStorage.setItem("token", response.data.token)
-                                        navigate("/home")
+                                        console.log(response)
+                                        if (response.data.msg == "User does not exist") {
+                                            alert("User does not exist Signup")
+                                        }
+                                        else if (response.data.msg == "Invalid Inputs") {
+                                            alert("Invalid Inputs")
+                                        }
+                                        else {
+                                            window.localStorage.setItem("token", response.data.token)
+                                            navigate("/home")
+                                        }
                                     } catch (error) {
                                         console.error("Error signing in form:", error);
                                     }
@@ -64,6 +74,14 @@ export default function Signin() {
                         >
                             Submit
                         </button>
+                    </div>
+                    <div className="py-2 flex justify-center">
+                        <div>
+                            Don't have account
+                        </div>
+                        <div className="underline pl-2" onClick={() => navigate("/signup")}>
+                            Signup
+                        </div>
                     </div>
                 </div>
             </div>
